+++ examples/create-logs.ts
@@ -1,0 +1,7 @@
+import { logger } from "../src/logger.js";
+
+logger.debug('This is a debug log');
+logger.info('This is an info log');
+logger.warn('This is a warn log');
+logger.error('This is an error log');
+logger.fatal('This is a fatal log');