+++ src/domain/entities/log.entity.ts
@@ -1,8 +1,4 @@
-export enum LogSeverity {
-    info = 'info',
-    warn = 'warn',
-    error = 'error'
-};
+import { LogSeverity } from "../enums/logSeverity.enum.js";
 
 export class LogEntity {
     message: string;
@@ -15,19 +11,23 @@ export class LogEntity {
         this.timestamp = log.timestamp ?? new Date();
     };
 
-    static fromJSON(content: string): LogEntity | void {
-        if (!content || content === '') return;
+    static fromJSON(content: string): LogEntity | void{
+        const log = JSON.parse(content);
 
-        const log: LogEntity = JSON.parse(content);
+        if (typeof log !== "object" || log === null) return;
 
-        if (
-            !log.message ||
-            !log.level ||
-            !log.timestamp
-        ) return;
+        if (typeof log.message !== 'string') return;
+
+        if (!Object.values(LogSeverity).includes(log.level)) return;
+
+        if (log.timestamp && typeof log.timestamp === 'string') {
+            const date = new Date(log.timestamp);
+            if (isNaN(date.getTime())) return;
+        } else return;
 
         const logEntity: LogEntity = new LogEntity(log);
-
         return logEntity;
     };
+
+
 };