+++ src/domain/enums/logSeverity.enum.ts
@@ -1,0 +1,7 @@
+export enum LogSeverity {
+    debug = 'debug',
+    info = 'info',
+    warn = 'warn',
+    error = 'error',
+    fatal = 'fatal'
+};